fix: renamed and edited the code of the .rs files to follow coding standards
