+++ src/code_generation/SimpleMacro.rs
@@ -1,10 +1,12 @@
-macro_rules! say_hello {
+macro_rules! say_hello
+    {
     () =>{
         println!("Hello World")
     }
-}
+    }
 
-pub fn main() {
+pub fn main()
+    {
     // Call the macro
     say_hello!();
-}
+    }
