+++ src/code_generation/Struct.rs
@@ -1,4 +1,5 @@
-struct User {
+struct User
+    {
     name: String,
     age: u32,
-}
+    }
