+++ src/code_generation/TraitsAndStructs.rs
@@ -1,16 +1,19 @@
-trait Vehicle {
+trait Vehicle
+    {
     fn mileage(&self) -> u32;
     fn fuel_type(&self) -> String;
-}
+    }
 
-struct Car {
+struct Car
+    {
     name: String,
     manufacturer: String,
-}
+    }
 
-struct Airplane {
+struct Airplane
+    {
     manufacturer: String,
     plane_type: String,
-}
+    }
 
 //todo: implement Vehicle for Car/Airplane
